pyproject.toml
[project]
name = "lama-triton"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.133.1",
    "gdown>=5.2.1",
    "numpy>=2.4.2",
    "onnx>=1.20.1",
    "onnxruntime-gpu>=1.24.2",
    "onnxscript>=0.6.2",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "tensorrt>=10.15.1.29",
    "torch<2.5",
    "uvloop>=0.21.0",
    "tritonclient[all]>=2.65.0",
    "uvicorn>=0.41.0",
    "psutil>=7.2.2",
]
